cell_area_measurement.ijm
After live/dead imaging of spheroids on the confocal and 2-photon, I got a z-stack series of images from both channels. I exported images as greyscale, and eventually in the folder I had tiff files with the names ending by ch00 (for green) and ch01 (for red). Then, I wanted to calculate the total area of green (live) and red (dead) cells in each layer of my z-stack to be able to calculate the ratio afterwards.